requirements.txt
imgaug==0.4.0
matplotlib==3.2.2
numpy==1.19.0
opencv-contrib-python==4.4.0.40
pandas==1.0.5
Pillow==7.2.0
scikit-learn==0.23.1
tensorboard==2.2.2
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-datasets==3.2.1
